cli: move the node shutdown code to its own file
Release note: None
